api routes protected

+++ backend/routes/todos.js
@@ -1,22 +1,22 @@
 import { Todo } from '../models/todo.js'
-//import auth from '../middleware/auth.js'
+import auth from '../middleware/auth.js'
 import express from 'express'
 import Joi from 'joi'
 
 const router = express.Router()
 
-router.get('/', async (req, res) => {
+router.get('/', auth, async (req, res) => {
     try {
         const todos = await Todo.find().sort({ date: -1 })
-
-        res.send(todos)
+        const filteredTodos = todos.filter(todo => todo.uid === req.user._id)
+        res.send(filteredTodos)
     } catch (err) {
         res.status(500).send(err.message)
         console.log(err.message)
     }
 });
 
-router.post('/', async (req, res) => {
+router.post('/', auth, async (req, res) => {
     const schema = Joi.object({
         name: Joi.string().min(3).max(200).required(),
         author: Joi.string().min(3).max(30),
@@ -48,7 +48,7 @@ router.post('/', async (req, res) => {
     }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
     const schema = Joi.object({
         name: Joi.string().min(3).max(200).required(),
         author: Joi.string().min(3).max(30),
@@ -69,6 +69,8 @@ router.put("/:id", async (req, res) => {
             return res.status(404).send("Todo not found.")
         }
 
+        if (todo.uid !== req.user._id) return res.status(401).send("Todo update failed. Not authorized.")
+
         const { name, author, uid, isComplete, date } = req.body
 
         const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, { name, author, uid, isComplete, date }, { new: true })
@@ -81,11 +83,15 @@ router.put("/:id", async (req, res) => {
     }
 });
 
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
     try {
         const todo = await Todo.findById(req.params.id)
 
         if (!todo) return res.status(404).send("Todo not found.")
+
+        if (todo.uid !== req.user._id) {
+            return res.status(401).send("Todo status change failed. Not authorized")
+        }
 
         const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, { isComplete: !todo.isComplete }, { new: true })
 
@@ -102,6 +108,10 @@ router.delete("/:id", async (req, res) => {
         const todo = await Todo.findById(req.params.id)
 
         if (!todo) return res.status(404).send("Todo not found.")
+
+        if (todo.uid !== req.user._id) {
+            return res.status(401).send("Todo deletion failed. Not authorized")
+        }
 
         const deletedTodo = await Todo.findByIdAndDelete(req.params.id)
